DdLogs integrated on Android example app
