+++ packages/core/android/src/main/kotlin/com/datadog/reactnative/GlobalState.kt
@@ -2,10 +2,10 @@ package com.datadog.reactnative
 
 import java.util.concurrent.ConcurrentHashMap
 
-internal object GlobalState {
+object GlobalState {
 
     // Keeps a record of global attributes for logs / spans
-    internal val globalAttributes: MutableMap<String, Any?> = ConcurrentHashMap()
+    val globalAttributes: MutableMap<String, Any?> = ConcurrentHashMap()
 
     /**
      * Adds a global attribute to all future RUM events.
