+++ packages/logs/android/gradle.properties
@@ -1,0 +1,6 @@
+DatadogSdkReactNativeLogs_kotlinVersion=1.8.21
+DatadogSdkReactNativeLogs_minSdkVersion=21
+DatadogSdkReactNativeLogs_compileSdkVersion=33
+DatadogSdkReactNativeLogs_buildToolsVersion=33.0.0
+DatadogSdkReactNativeLogs_targetSdkVersion=33
+android.useAndroidX=true
