+++ packages/logs/android/src/main/AndroidManifest.xml
@@ -1,0 +1,11 @@
+<!--
+Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+This product includes software developed at Datadog (https://www.datadoghq.com/).
+Copyright 2016-Present Datadog, Inc.
+-->
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          package="com.datadog.reactnative.logs">
+
+    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+</manifest>
