+++ packages/logs/android/src/newarch/kotlin/com/datadog/reactnative/logs/DdLogs.kt
@@ -4,11 +4,10 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.reactnative
+package com.datadog.reactnative.logs
 
 import com.facebook.react.bridge.Promise
 import com.facebook.react.bridge.ReactApplicationContext
-import com.facebook.react.bridge.ReactContextBaseJavaModule
 import com.facebook.react.bridge.ReactMethod
 import com.facebook.react.bridge.ReadableMap
 
@@ -18,7 +17,7 @@ import com.facebook.react.bridge.ReadableMap
 class DdLogs(
     reactContext: ReactApplicationContext,
     datadogWrapper: DatadogWrapper = DatadogSDKWrapper()
-) : ReactContextBaseJavaModule(reactContext) {
+) : NativeDdLogsSpec(reactContext) {
 
     private val implementation = DdLogsImplementation(datadog = datadogWrapper)
 
@@ -29,7 +28,7 @@ class DdLogs(
      * @param configuration The logs configuration.
      */
     @ReactMethod
-    fun enable(configuration: ReadableMap, promise: Promise) {
+    override fun enable(configuration: ReadableMap, promise: Promise) {
         implementation.enable(configuration, promise)
     }
 
@@ -39,7 +38,7 @@ class DdLogs(
      * @param context The additional context to send.
      */
     @ReactMethod
-    fun debug(message: String, context: ReadableMap, promise: Promise) {
+    override fun debug(message: String, context: ReadableMap, promise: Promise) {
         implementation.debug(message, context, promise)
     }
 
@@ -49,7 +48,7 @@ class DdLogs(
      * @param context The additional context to send.
      */
     @ReactMethod
-    fun info(message: String, context: ReadableMap, promise: Promise) {
+    override fun info(message: String, context: ReadableMap, promise: Promise) {
         implementation.info(message, context, promise)
     }
 
@@ -59,7 +58,7 @@ class DdLogs(
      * @param context The additional context to send.
      */
     @ReactMethod
-    fun warn(message: String, context: ReadableMap, promise: Promise) {
+    override fun warn(message: String, context: ReadableMap, promise: Promise) {
         implementation.warn(message, context, promise)
     }
 
@@ -69,7 +68,7 @@ class DdLogs(
      * @param context The additional context to send.
      */
     @ReactMethod
-    fun error(message: String, context: ReadableMap, promise: Promise) {
+    override fun error(message: String, context: ReadableMap, promise: Promise) {
         implementation.error(message, context, promise)
     }
 
@@ -83,7 +82,7 @@ class DdLogs(
      */
     @Suppress("LongParameterList")
     @ReactMethod
-    fun debugWithError(
+    override fun debugWithError(
         message: String,
         errorKind: String?,
         errorMessage: String?,
@@ -91,14 +90,7 @@ class DdLogs(
         context: ReadableMap,
         promise: Promise
     ) {
-        implementation.debugWithError(
-            message,
-            errorKind,
-            errorMessage,
-            stacktrace,
-            context,
-            promise
-        )
+        implementation.debugWithError(message, errorKind, errorMessage, stacktrace, context, promise)
     }
 
     /**
@@ -111,7 +103,7 @@ class DdLogs(
      */
     @Suppress("LongParameterList")
     @ReactMethod
-    fun infoWithError(
+    override fun infoWithError(
         message: String,
         errorKind: String?,
         errorMessage: String?,
@@ -132,7 +124,7 @@ class DdLogs(
      */
     @Suppress("LongParameterList")
     @ReactMethod
-    fun warnWithError(
+    override fun warnWithError(
         message: String,
         errorKind: String?,
         errorMessage: String?,
@@ -153,7 +145,7 @@ class DdLogs(
      */
     @Suppress("LongParameterList")
     @ReactMethod
-    fun errorWithError(
+    override fun errorWithError(
         message: String,
         errorKind: String?,
         errorMessage: String?,
@@ -161,13 +153,6 @@ class DdLogs(
         context: ReadableMap,
         promise: Promise
     ) {
-        implementation.errorWithError(
-            message,
-            errorKind,
-            errorMessage,
-            stacktrace,
-            context,
-            promise
-        )
+        implementation.errorWithError(message, errorKind, errorMessage, stacktrace, context, promise)
     }
 }
