+++ packages/logs/tsconfig.json
@@ -1,3 +1,3 @@
 {
-  "extends": "../../tsconfig"
+  "extends": "../../tsconfig",
 }
